requirements.txt
aiogram==3.13.1
aiofiles==23.2.1
python-dotenv==1.0.1
asyncio-mqtt==0.16.2